ai_room_cleaner/pyproject.toml
[project]
name = "ai-room-cleaner"
version = "0.1.0"
description = "AI-powered room cleanliness analysis tool."
authors = [
    {name = "AI Software Engineer"}
]
readme = "../README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "bleach (>=6.2.0,<7.0.0)",
    "pyvips (>=3.0.0,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "openai (>=1.86.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "werkzeug (>=3.0.3,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "cachetools (>=5.3.3,<6.0.0)"
]

[tool.poetry]
packages = [
    { include = "backend", from = "." },
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^5.0.0"
pip-audit = "^2.7.0"
pylint = "^3.2.5"

[tool.pylint.main]
ignore-patterns = ["^\\..*", "dist", "node_modules", "venv"]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
